Breaks the circular dependency
